% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggzcql-package.R
\docType{package}
\name{ggzcql-package}
\alias{ggzcql}
\alias{ggzcql-package}
\title{ggzcql: 'ggzcql': Color Palettes for 'ggplot2'}
\description{
A collection of color palettes for 'ggplot2' following the design principles of 'ggsci'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nanxstats/ggzcql}
  \item Report bugs at \url{https://github.com/nanxstats/ggzcql/issues}
}

}
\author{
\strong{Maintainer}: Kevin Wang \email{kevin.wang@example.com}

}
\keyword{internal}
