% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcql.R
\name{pal_zcql}
\alias{pal_zcql}
\title{zcql Palette}
\usage{
pal_zcql(palette = c("default"), alpha = 1)
}
\arguments{
\item{palette}{Palette type ("default").}

\item{alpha}{Transparency level, a number between 0 and 1.}
}
\description{
Color palette inspired by "A Thousand Li of Rivers and Mountains" (千里江山图).
}
\examples{
library("scales")
show_col(pal_zcql()(5))
}
