% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcql.R
\name{scale_color_zcql}
\alias{scale_color_zcql}
\alias{scale_colour_zcql}
\alias{scale_fill_zcql}
\title{zcql Color Scales for ggplot2}
\usage{
scale_color_zcql(palette = c("default"), alpha = 1, ...)

scale_colour_zcql(palette = c("default"), alpha = 1, ...)

scale_fill_zcql(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type ("default").}

\item{alpha}{Transparency level, a number between 0 and 1.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_zcql]{pal_zcql()}} for details.
}
\examples{
library("ggplot2")
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  scale_color_zcql() +
  theme_zcql()
}
