% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcql.R
\name{theme_zcql}
\alias{theme_zcql}
\title{zcql Theme for ggplot2}
\usage{
theme_zcql(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\description{
A custom ggplot2 theme with a light golden silk background (\verb{#E8D7B0}) inspired by the luminous silk canvas of "A Thousand Li of Rivers and Mountains".
}
\examples{
library("ggplot2")
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  scale_color_zcql() +
  theme_zcql()
}
